Implement Sqlite db, and in memory Sqlite Fix DLA concreate classes for db sets Fix model for concreate interface impl

+++ Model.Common/src/IVehicleRegistration.cs
@@ -3,9 +3,9 @@
 public interface IVehicleRegistration : IBaseEntity
 {
     string RegistrationNumber { get; set; }
-    int VehicleModelId { get; set; }
-    int VehicleEngineTypeId { get; set; }
-    int VehicleOwnerId { get; set; }
+    long VehicleModelId { get; set; }
+    long VehicleEngineTypeId { get; set; }
+    long VehicleOwnerId { get; set; }
     IVehicleModel Model { get; set; }
     IVehicleEngineType EngineType { get; set; }
     IVehicleOwner Owner { get; set; }
